Areas completo

+++ app/src/main/java/com/example/android/listview/Metodos.java
@@ -2,19 +2,19 @@ package com.example.android.listview;
 
 public class Metodos {
 
-    public static int areaCuadrado(int lado) {
+    public static double areaCuadrado(double lado) {
         return lado*lado;
     }
 
-    public static int areaRectangulo(int base, int altura) {
+    public static double areaRectangulo(double base, double altura) {
         return base * altura;
     }
 
-    public static int areaTriangulo(int base, int altura) {
+    public static double areaTriangulo(double base, double altura) {
         return (base * altura) / 2;
     }
 
-    public static double areaCirculo(int radio){
+    public static double areaCirculo(double radio){
         return radio * radio * Math.PI;
     }
